Metadata fetching process removed from repo to AudioMetadataManager
